+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Backsplash server does its own access logging [#5531](https://github.com/raster-foundry/raster-foundry/pull/5531)
 - Include annotation project id + TMS triple in traces for MVT requests [#5533](https://github.com/raster-foundry/raster-foundry/pull/5533)
 - Added label class and label class group endpoints for campaigns [#5534](https://github.com/raster-foundry/raster-foundry/pull/5534)
+- Added a QP for granting requester accesses to project and campaign in the campaign random task GET endpoint [#5539](https://github.com/raster-foundry/raster-foundry/pull/5539)
 
 ### Removed
 - Remove Nginx configuration and containers for Backsplash [#5532](https://github.com/raster-foundry/raster-foundry/pull/5532)